+++ src/py/flwr/server/strategy/fedavg_android.py
@@ -40,19 +40,6 @@ from flwr.server.client_proxy import ClientProxy
 from .aggregate import aggregate, weighted_loss_avg
 from .strategy import Strategy
 
-DEPRECATION_WARNING = """
-DEPRECATION WARNING: deprecated `eval_fn` return format
-
-    loss, accuracy
-
-move to
-
-    loss, {"accuracy": accuracy}
-
-instead. Note that compatibility with the deprecated return format will be
-removed in a future release.
-"""
-
 DEPRECATION_WARNING_INITIAL_PARAMETERS = """
 DEPRECATION WARNING: deprecated initial parameter type
 
@@ -104,8 +91,8 @@ class FedAvgAndroid(Strategy):
                 during validation. Defaults to 2.
             min_available_clients (int, optional): Minimum number of total
                 clients in the system. Defaults to 2.
-            eval_fn (Callable[[Weights], Optional[Tuple[float, float]]], optional):
-                Function used for validation. Defaults to None.
+            eval_fn : Callable[[Weights], Optional[Tuple[float, Dict[str, Scalar]]]]
+                Optional function used for validation. Defaults to None.
             on_fit_config_fn (Callable[[int], Dict[str, Scalar]], optional):
                 Function used to configure training. Defaults to None.
             on_evaluate_config_fn (Callable[[int], Dict[str, Scalar]], optional):
@@ -163,12 +150,7 @@ class FedAvgAndroid(Strategy):
         eval_res = self.eval_fn(weights)
         if eval_res is None:
             return None
-        loss, other = eval_res
-        if isinstance(other, float):
-            print(DEPRECATION_WARNING)
-            metrics = {"accuracy": other}
-        else:
-            metrics = other
+        loss, metrics = eval_res
         return loss, metrics
 
     def configure_fit(